Move command extraction logic to utils
